Move demo back to root, and adjust webpack config to suit
