+++ demo/collab/SelectionPlugin.ts
